Fixed a problem in getTotalRows() which caluclates the total number of rows by setting the current cusor to the last row and getting the row number. After setting the cursor to the last row, the cursor is set to the first row. For cases where the cursor was originally  before the first row, this resulted in losing the first row when next() is called.
The solution was to save the current row position and returning the cursor back to the original
position if the originl cursor was on a row already. If the original cursor position was
not on a row (before first or after last), then the DEFAULT BEHAVIOR right now
is to set the cursor to before first.