Refactored   so that it is easier to use each field's own flags and states with their respective field class.
Also passing in the field to  the constructStateInstance() method of the field class. This is so that a state object is always associated with a field and if we do move the state out of the field class, it will be easier to refactor.
Changed the getField() method of each State inner class to return the actual passed in field object instead of returning the outer class object.