Converted the old HtmlTabularReportDataSourceScrollStates concrete class into an interface and created two concrete classes that implement that interface using two different strategies.
The first strategy (the default one) stores scroll states in the user's HTTP session object. This strategy is simple and and useful in a non-clustered envrionment where the session weight is not important.

The second strategy stores scroll states outside of a HTTP session. This strategy is more useful in a clustered environment where the session weight is important (because scroll states are no serializable because they hang onto a database connection).