Added capability to create a <insert-additional-data> template producer that would be executed whenever inserts were done. Example:
<dialog type="table-dialog" ...>
    ... bind table normally ...

    <insert-additional-data>
        <schema.table col1="abc" col2="request:def" col3="xyz"/>
        <table _conditional="expression" col1="abc" col2="request:def"/>
    </insert-additional-data>

In the above example, the normal table dialog inserts (according to bind values) would take place and then the table dialog would start to insert the data in the tables provided in <insert-additional-data>

<schema.table> or <table> may be supplied (if just a table is supplied, then the bind-table schema is used or the first available schema if bind-table is not present)

the col="XXX" sets the column values to either a value source if XXX is a value source or to the field XXX's value if XXX is not a value source.

the _conditional="expr" allows execution of an expression and the insert only takes place if the expr returns true. The expression is a JSTL-style expression and the variable "vc" (the value context -- which is really a TableDialogContext instance) is available for use in dynamic expressions.