Added a simple but potentially very important change to the way that navigation pages handle dialogs and other commands that could affect navigation (like redirecting to another page).  PLEASE TEST YOUR APPS THOROUGHLY TO ENSURE NOTHING IS BROKEN.
Using the new HttpServletCommand.isAbleToAffectNavigation() method, the page now automatically sets the BODY_AFFECTS_NAVIGATION flag if its appropriate.

Later, when rendering the page if the body may affect the navigation (through the use of the BODY_AFFECTS_NAVIGATION flag) the navigation page can now check to see if a redirect has occurred and not perform any meta data, header, and footer rendering. This is done by executing the dialog (or other command) into a string writer and then checking to see if a redirect occurred.

If a redirect does occur, then the body is discarded and the meta data, header, and footer skin calls will not be made.

This will actually speed up performance considerably in many cases where the execution of a dialog simply redirects to another page after conducting its business. The downside is that for cases where the dialog is not doing a redirect the body will be stored in a tempory variable (string writer) and then emit the contents to the response buffer. The temporary variable may end up eating up a bit more memory but the command instance can optimize that if required.