Added capability for DAL import XML files to import CSV delimited text entries. Introduced two new 'dal:csv' and 'dal:csv-file' attributes that perform the action. For example, in the 'initial-and-test-data.xml' you can now do the following:
<dal xmlns:dal="http://www.netspective.org/Framework/Axiom/DataAccessLayer">
    <book-info dal:csv="test-cvs-01,CSV Book 01,&quot;Shah, Shahid N.&quot;,Mystery,ISBNCSV01"/>
    <book-info dal:csv-file="c:\X\Y\initial-and-test-data.csv"/>
</dal>

The first version (dal:csv) supports a single line of CSV. The second points to a file. In either version, all columns must be provided in the order they were defined in the Schema XML.

Be sure to add the xmlns:dal="http://www.netspective.org/Framework/Axiom/DataAccessLayer" namespace in the <dal> tag. And, just just your table name in place of <book-info>.