Added ability for values to generate URL params. This feature is used by dialog fields' new activity recording feature where instead of storing the dialog values in XML, they store them auto-executable URL params.
Added new activity package that will allow generic events (called activities) to be broadcasted and observed. The initial purpose of this feature is to allow navigation pages and dialogs to announce that they've been visited or executed and allow observers to record the visit or execution (so that they can show a "recently visisted" history page or let dialogs be auto-executed via single click).