Added experimental support for "dynamic" comparisons in query defnitions.
This is useful when the actual comparison may not be known until a user selects a field or chooses a comparison dynamically. Now, in the <condition> tag of the <select> dialog in a <query-defn>, the following can be specified:

<condition field="x" ... comparison="dynamic,value:source"...>

If a comparison called "dynamic" is seen, a value source must follow a comma. Then, at runtime the dynamic comparison will "run" the value source, get back a real comparison id then use the comparison ID that was generated dynamically by the value source.