Major refactoring of AuthenticatedOrgUser interface.
* Removed the AuthenticatedOrgUser and AuthenticatedOrgsUser interfaces (plus their mutable counterparts) and implementations
* Added new AuthenticatedOrganization and AuthenticatedOrganizations interfaces (plus their mutable counterparts) and implementations
* Added new EntityPreference and EntityPreferences interfaces and implementations.

Now, the AuthenticatedUser includes a new method getOrganizations() which is a AuthenticatedOrganizations implementation. This method always returns a non-null value so if a user doesn't have any organizations, it's just empty. So, the auth user can own one more organizations with full knowlege of a "primary" organization.

Also added a new AuthenticatedUser.getPreferences() method that allows users to manage their own preferences. Most of the time the preferences are read from a database and these preference classes have nothing to do with the JDK 1.4 preferences object. Orgs also have a preferences method that allows orgs to manage general preferences.