CODE MAY BREAK: Changed the way JDNI connection provider works.
Now, instead of using local JNDI environment by default, the connection provider uses a root environment and allows the root context name to be supplied via the XDM.

If you want to use the old behvaior, you must use the following in your project.xml file:

<connection-provider class="com.netspective.xom.connection.JndiConnectionProvider" root-context-name="java:/comp/env/jdbc"/>

If you don't supply the local root context, app servers like Tomcat or Resin may not work properly.

This change was made to support clustering of JDBC connection pools on app servers like WebLogic and WebSphere.

Another change that was made was that we moved away from static InitialContext to instance field. This is probably more appropriate so that root context could be overriden.