Add new "Entity Variant" table type and associated Java classes.
An entity variant table allows an RDBMS to manage "variant" attributes for a particular entity. 
For example, if a Person table needs to manage an unlimited number of "soft attribute" meaning
data about a person that is not managed in the primary table this variant record manager will
do the trick.

As an example, the following is a definition of an "App User Preferences" object: 

<table name="App_User_Pref_Type" type="Entity_Variant_Record_Type" abbrev="A_U_PF_T"
   xdm:param-tableDescr="The types of preferences that a user can specify"
   xdm:param-variantColNamePrefix="pref">

    <types>
        <type pref_type_name="Selected Searched Products Sort Order" pref_type_java_class="java.lang.String" pref_type_default="p.item_name"/>
        <type pref_type_name="Show Selected Searched Product Images" pref_type_java_class="java.lang.Boolean" pref_type_default="false"/>
        <type pref_type_name="Selected Catalog Products Sort Order" pref_type_java_class="java.lang.String" pref_type_default="p.item_name"/>
        <type pref_type_name="Show Selected Catalog Product Images" pref_type_java_class="java.lang.Boolean" pref_type_default="false"/>
    </types>
</table>

<table name="App_User_Pref" type="Default,Entity_Variant_Record" abbrev="A_U_PF"
       xdm:param-tableDescr="A User's preferences"
       xdm:param-variantColNamePrefix="pref"
       xdm:param-variantTypeTable="App_User_Pref_Type"
       xdm:param-entityRefColName="user_id"
       xdm:param-entityRef="App_User">

   <index name="A_U_PF_unq_pref" unique="yes" columns="user_id,pref_type_id"/>
</table>

These two tables together will allow as many "preference types" to be added to the type
table. If new application preferences are required they just become new rows in the
type table and nothing else needs to change.

When you need to read the values in from the database, the custom table classes have 
the ability to read all preferences into a map using the following DAL code:

AppUserPrefTable.Records records = null;
records = appUserRecord.getAppUserPrefTableRecords(cc);
userPreferences = records.createNameValueMap();

Although there are other methods available for managing the special types and variant
values tables, the createNameValueMap() method is the most convenient. It lets you
read all the records for a particular parent (appUserRecord.getAppUserPrefTableRecords(cc))
and transfer them to a Map. 

Then, using DAL-based Java triggers you can reread the records and keep a cache. This is
new feature and is still subject to change but it can handle a good deal of complexity and
put it into Axiom for meta-model based data models.