Add a new insertOrUpdateIfDuplicateRowFound() method to Table.
This method that allows a record to be automatically inserted or updated based on
whether the record exists in the database already. The first for duplicate records
is done is by primary key and the second check is by unique indexes. If a dupe
is found, it is assumed that we want it replaced. If no dupes are found (either
by primary key or by unique indexes) then the record is just inserted.