Improved support for multiple <sql> blocks within a <query> tag. Within the code, moved the QueryParameters so that now it is owned by the <sql> tag instead of <query>.
This allows the following now:

<query name="x">
   select * from x where y = ?
   <params>
      ...
   </params>
</query>

The above example is the "standard" example which now implicitly do the following translation:

<query name="x">
   <sql dbms="ansi">
       select * from x where y = ?
       <params>
          ...
       </params>
   </sql>
</query>

Before the recent change, the <params> tag always belonged to the <query> tag. This means that even if you had multiple SQL statements for a <query> you could only have one set of parameters. This has now been fixed so that each <sql> tag contains and allows its own parameters.

This change is also required to implement the upcoming role-based query system which will allow different <sql> blocks to be provided for different roles of a user.