Added capability for index directory to be specified in one of three ways
1) index-directory attribute (for a known location)

2) index-locator-properties with properties in a separate file (located using app's classpath)

3) <index-directory-search-path>
      <directory path="xxx">

The second and third models are both quite flexible in that they allow multiple directories to be supplied and the first directory that was found will be used.

This allows a set of directories to be standardized across multiple operating systems and the first directory that is valid will be used. 
<index-directory-search-path>
    <directory path="../../../fts/products">
    <directory path="C:\FTS\products">
    <directory path="/etc/var/products">
    <directory path="/some/other/loc">
</index-directory>

In the above example the first directory would be relative to the <page> tag where the FTS index page was defined. The next paths would be searched in order and would match the first valid directory. Once a path is found, the rest are not searched.

The same works with the properties file except the property String can be delimited with ',' to separate multiple search paths like so:

indexDirPathSep=,
indexDir=../../../fts/products, C:\FTS\products, ...etc...