Fixed bug where a dialog that was executed appropriately is re-executed using the back button but the second execution results in a validation error and the dialog executed anyway.
Problem was researched by Carol Dionne of CardinalHealth in this mannner (found in SupplyCentral bug 328):

Source of problem is in NEFS' Dialog class when a "transaction" page is 
resubmitted via the back button. The DialogState 'executeMode' of the 
DialogContext is not set properly during the processing of the resubmit so even 
though the validations find errors (and uses addError) the request is executed 
instead of sending back an error page.

Here is the gist of the processing flow to explain source of problem:
Submit the Add User page. 
The add is successful, the DialogContext.DialogState has executeMode = true
Use the back button in the browser to return to the add user page
Click on "save" again.
The DialogContext.initialize method pulls the DialogState from the http request 
with executeMode = true. Note, on the first submit, the executeMode = false at 
this point.


Processing flows into Dialog.prepareContext which calls DialogContext.calcState
and then checks the dialogState to see if page should be executed. 
DialogContext.calcState which runs the validations for the Add User dialog 
page. The validations in UserDialogValidationListener correctly detect a 
duplicate user name and use AddError to flag the error. CalcState sets 
executeMode to true only if no validation errors. It does not set executeMode 
to false if errors are found. CalcState returns false. At this point on the 
resubmit, executeMode is still true. On the first submit, executeMode remains 
false.

Control returns to Dialog.prepareContext which tests the dialogContext's execu 
DialogState.executeMode; since its true (incorrectly) it goes ahead and 
executes the request which leads to the database error.

A fix is needed for resubmitted pages to get the executeMode set to false if 
validation errors are found. Possible fix is to add else clause to the if 
statement at line 685 in  DialogContext.calcState to set executeMode to false.